linux an nonlinux behavior for tcp
